fix(obsidian): strip Windows verbatim path prefix from tool outputs
Path::canonicalize on Windows produces \?\-prefixed verbatim paths
that leak into agent tool JSON output (list_files, read_file, etc.).
The AI model sees confusing paths like \?\C:\Users\... which degrades
tool-call quality and may contribute to upstream provider failures.

Move normalize_vault_path_for_external_use from obsidian.rs into
filesystem.rs as the shared normalize_path_for_external_use helper
taking &Path instead of consuming PathBuf. Apply it at every
model-facing output boundary (JSON path fields, search match lines,
generated media paths) while keeping the raw canonical path for
internal containment checks and filesystem operations.
